requirements.txt
libtorrent
nuitka
requests
requests-toolbelt
gofilepy-api
humanize